move `TableId` to `common`
