+++ crates/store/re_grpc_client/src/message_proxy/write.rs
@@ -6,8 +6,8 @@ use re_log_encoding::codec::wire::encoder::Encode as _;
 use re_log_encoding::Compression;
 use re_log_types::LogMsg;
 use re_log_types::TableMsg;
+use re_protos::common::v1alpha1::TableId as TableIdProto;
 use re_protos::sdk_comms::v1alpha1::message_proxy_service_client::MessageProxyServiceClient;
-use re_protos::sdk_comms::v1alpha1::TableId as TableIdProto;
 use re_protos::sdk_comms::v1alpha1::WriteMessagesRequest;
 use re_protos::sdk_comms::v1alpha1::WriteTableRequest;
 use re_uri::ProxyEndpoint;
