+++ crates/store/re_protos/proto/rerun/v1alpha1/common.proto
@@ -28,6 +28,11 @@ message RecordingId {
   string id = 1;
 }
 
+// uniquely identifies a table
+message TableId {
+  string id = 1;
+}
+
 // A recording can have multiple timelines, each is identified by a name, for example `log_tick`, `log_time`, etc.
 message Timeline {
   string name = 1;
