+++ crates/store/re_protos/src/v1alpha1/rerun.common.v1alpha1.rs
@@ -38,6 +38,22 @@ impl ::prost::Name for RecordingId {
         "/rerun.common.v1alpha1.RecordingId".into()
     }
 }
+/// uniquely identifies a table
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct TableId {
+    #[prost(string, tag = "1")]
+    pub id: ::prost::alloc::string::String,
+}
+impl ::prost::Name for TableId {
+    const NAME: &'static str = "TableId";
+    const PACKAGE: &'static str = "rerun.common.v1alpha1";
+    fn full_name() -> ::prost::alloc::string::String {
+        "rerun.common.v1alpha1.TableId".into()
+    }
+    fn type_url() -> ::prost::alloc::string::String {
+        "/rerun.common.v1alpha1.TableId".into()
+    }
+}
 /// A recording can have multiple timelines, each is identified by a name, for example `log_tick`, `log_time`, etc.
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Timeline {
